+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/account/AccountNotFoundException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.purchase.exception.account;
 
 public class AccountNotFoundException extends RuntimeException{
   
-  public AccountNotFoundException(int accountId, Throwable ex) {
+  public AccountNotFoundException(long accountId, Throwable ex) {
     super("The account " + accountId + " wasn't found.", ex);
   }
