+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/account/AccountRepositoryException.java
@@ -1,7 +1,9 @@
 package com.jpdr.apps.demo.webflux.purchase.exception.account;
 
 public class AccountRepositoryException extends RuntimeException{
-  public AccountRepositoryException(int accountId, Throwable ex) {
+  
+  public AccountRepositoryException(long accountId, Throwable ex) {
     super("There was an error while retrieving the account " + accountId + ".", ex);
   }
+  
 }
