+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/stock/StockNotFoundException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.purchase.exception.stock;
 
 public class StockNotFoundException extends RuntimeException{
   
-  public StockNotFoundException(Integer productId, Throwable ex){
+  public StockNotFoundException(Long productId, Throwable ex){
     super("There is no stock for product " + productId, ex);
   }
