+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/stock/StockRepositoryException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.purchase.exception.stock;
 
 public class StockRepositoryException extends RuntimeException{
   
-  public StockRepositoryException(Integer productId, Throwable ex){
+  public StockRepositoryException(Long productId, Throwable ex){
     super("An error occurred while retrieving stock data for product " + productId +".",ex);
   }
