+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/user/UserNotFoundException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.purchase.exception.user;
 
 public class UserNotFoundException extends RuntimeException{
   
-  public UserNotFoundException(int userId, Throwable ex){
+  public UserNotFoundException(long userId, Throwable ex){
     super("User "+ userId +" not found", ex);
   }
