+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/exception/user/UserRepositoryException.java
@@ -2,7 +2,7 @@ package com.jpdr.apps.demo.webflux.purchase.exception.user;
 
 public class UserRepositoryException extends RuntimeException{
   
-  public UserRepositoryException(Integer userId, Throwable ex){
+  public UserRepositoryException(Long userId, Throwable ex){
     super("An error occurred while retrieving data for user " + userId +".",ex);
   }
