+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/repository/account/AccountRepository.java
@@ -5,8 +5,8 @@ import com.jpdr.apps.demo.webflux.purchase.service.dto.account.AccountDto;
 import com.jpdr.apps.demo.webflux.purchase.service.dto.account.AccountTransactionDto;
 import reactor.core.publisher.Mono;
 
-public interface AccountRepository extends BaseRepository<AccountDto, Integer> {
+public interface AccountRepository extends BaseRepository<AccountDto, Long> {
 
-  Mono<AccountTransactionDto> createTransaction(Integer accountId, AccountTransactionDto transaction);
+  Mono<AccountTransactionDto> createTransaction(Long accountId, AccountTransactionDto transaction);
 
 }
