+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/repository/product/ProductRepository.java
@@ -3,5 +3,5 @@ package com.jpdr.apps.demo.webflux.purchase.repository.product;
 import com.jpdr.apps.demo.webflux.purchase.repository.BaseRepository;
 import com.jpdr.apps.demo.webflux.purchase.service.dto.product.ProductDto;
 
-public interface ProductRepository extends BaseRepository<ProductDto, Integer> {
+public interface ProductRepository extends BaseRepository<ProductDto, Long> {
 }
