+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/repository/product/impl/ProductRepositoryImpl.java
@@ -27,7 +27,7 @@ public class ProductRepositoryImpl implements ProductRepository {
   
   @Override
   @Cacheable(key = "#productId", value = "products", sync = true)
-  public Mono<ProductDto> getById (Integer productId) {
+  public Mono<ProductDto> getById (Long productId) {
     return this.webClient.get()
       .uri("/products/{productId}", productId)
       .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
