+++ src/main/java/com/jpdr/apps/demo/webflux/purchase/repository/user/UserRepository.java
@@ -3,5 +3,5 @@ package com.jpdr.apps.demo.webflux.purchase.repository.user;
 import com.jpdr.apps.demo.webflux.purchase.repository.BaseRepository;
 import com.jpdr.apps.demo.webflux.purchase.service.dto.user.UserDto;
 
-public interface UserRepository extends BaseRepository<UserDto, Integer> {
+public interface UserRepository extends BaseRepository<UserDto, Long> {
 }
